move vacollect from octarender.cpp > vacollect.cpp
