+++ src/engine/render/vacollect.h
@@ -1,0 +1,6 @@
+#ifndef VACOLLECT_H_
+#define VACOLLECT_H_
+
+extern void destroyvbo(GLuint vbo);
+
+#endif
